# packages/settlrl-app/frontend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Settlrl Render",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/replay": {
   "post": {
    "summary": "Post Replay",
    "description": "Load a game record (the ``GameRecord`` JSON document) for replay;\nreturns the opening state. ``422`` if the record is malformed or fails\nreplay validation.",
    "operationId": "post_replay_api_replay_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "additionalProperties": true,
        "type": "object",
        "title": "Doc"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ReplayStateModel"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/replay/state": {
   "get": {
    "summary": "Get Replay State",
    "description": "The loaded replay after ``move`` moves (0 = the opening board).\n\n``404`` when no replay is loaded; ``422`` when ``move`` is out of\nrange.",
    "operationId": "get_replay_state_api_replay_state_get",
    "parameters": [
     {
      "name": "move",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 0,
       "title": "Move"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ReplayStateModel"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "BankModel": {
    "properties": {
     "resources": {
      "$ref": "#/components/schemas/ResourceCounts"
     },
     "dev_cards": {
      "type": "integer",
      "title": "Dev Cards"
     }
    },
    "type": "object",
    "required": [
     "resources",
     "dev_cards"
    ],
    "title": "BankModel",
    "description": "Cards left in the supply: resource stacks plus the development deck."
   },
...
   "BoardModel": {
    "properties": {
     "tiles": {
      "items": {
       "$ref": "#/components/schemas/TileModel"
      },
      "type": "array",
      "title": "Tiles"
     },
     "buildings": {
      "items": {
       "$ref": "#/components/schemas/BuildingModel"
      },
      "type": "array",
      "title": "Buildings",
      "default": []
     },
     "roads": {
      "items": {
       "$ref": "#/components/schemas/RoadModel"
      },
      "type": "array",
      "title": "Roads",
      "default": []
     },
     "ports": {
      "items": {
       "$ref": "#/components/schemas/PortModel"
      },
      "type": "array",
      "title": "Ports",
      "default": []
     },
     "players": {
      "items": {
       "$ref": "#/components/schemas/PlayerModel"
      },
      "type": "array",
      "title": "Players",
      "default": []
     },
     "robber": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/HexModel"
       },
       {
        "type": "null"
       }
      ]
     },
     "bank": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/BankModel"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "type": "object",
    "required": [
     "tiles"
    ],
    "title": "BoardModel"
   },
...
   "BuildingModel": {
    "properties": {
     "cube": {
      "$ref": "#/components/schemas/CubeModel"
     },
     "player": {
      "type": "integer",
      "title": "Player"
     },
     "kind": {
      "type": "string",
      "enum": [
       "settlement",
       "city"
      ],
      "title": "Kind"
     }
    },
    "type": "object",
    "required": [
     "cube",
     "player",
     "kind"
    ],
    "title": "BuildingModel",
    "description": "A settlement or city sitting on a vertex, owned by a 0-indexed player."
   },
   "CubeModel": {
    "properties": {
     "q": {
      "type": "integer",
      "title": "Q"
     },
     "r": {
      "type": "integer",
      "title": "R"
     },
     "s": {
      "type": "integer",
      "title": "S"
     }
    },
    "type": "object",
    "required": [
     "q",
     "r",
     "s"
    ],
    "title": "CubeModel",
    "description": "A board vertex, in cube coordinates (q + r + s = \u00b11)."
   },
   "DevCardCounts": {
    "properties": {
     "knight": {
      "type": "integer",
      "title": "Knight"
     },
     "road_building": {
      "type": "integer",
      "title": "Road Building"
     },
     "year_of_plenty": {
      "type": "integer",
      "title": "Year Of Plenty"
     },
     "monopoly": {
      "type": "integer",
      "title": "Monopoly"
     },
     "victory_point": {
      "type": "integer",
      "title": "Victory Point"
     }
    },
    "type": "object",
    "required": [
     "knight",
     "road_building",
     "year_of_plenty",
     "monopoly",
     "victory_point"
    ],
    "title": "DevCardCounts",
    "description": "Unplayed development cards in hand, by type (DevCard order)."
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "HexModel": {
    "properties": {
     "q": {
      "type": "integer",
      "title": "Q"
     },
     "r": {
      "type": "integer",
      "title": "R"
     }
    },
    "type": "object",
    "required": [
     "q",
     "r"
    ],
    "title": "HexModel",
    "description": "An axial (pointy-top) tile coordinate."
   },
   "LogEntryModel": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "kind": {
      "type": "string",
      "enum": [
       "move",
       "chat",
       "win"
      ],
      "title": "Kind"
     },
     "player": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Player"
     },
     "action_type": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Action Type"
     },
     "text": {
      "type": "string",
      "title": "Text",
      "default": ""
     }
    },
    "type": "object",
    "required": [
     "id",
     "kind"
    ],
    "title": "LogEntryModel",
    "description": "One line of the game's chat / log.\n\n``player`` is the seat the line belongs to (``None``: a spectator's chat\nmessage). Move entries carry the ``action_type`` (the client maps it to an\nicon) and a short text (the action label, or ``\"rolled N\"`` for rolls)."
   },
...
   "PlayerModel": {
    "properties": {
     "player": {
      "type": "integer",
      "title": "Player"
     },
     "resource_cards": {
      "type": "integer",
      "title": "Resource Cards"
     },
     "dev_cards": {
      "type": "integer",
      "title": "Dev Cards"
     },
     "victory_points": {
      "type": "integer",
      "title": "Victory Points"
     },
     "knights_played": {
      "type": "integer",
      "title": "Knights Played",
      "default": 0
     },
     "longest_road": {
      "type": "boolean",
      "title": "Longest Road",
      "default": false
     },
     "largest_army": {
      "type": "boolean",
      "title": "Largest Army",
      "default": false
     },
     "resources": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/ResourceCounts"
       },
       {
        "type": "null"
       }
      ]
     },
     "dev_card_types": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/DevCardCounts"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "type": "object",
    "required": [
     "player",
     "resource_cards",
     "dev_cards",
     "victory_points"
    ],
    "title": "PlayerModel",
    "description": "Summary stats for one (0-indexed) player.\n\nThe per-type breakdowns are private: the server nulls them out for every\nseat the requester doesn't own (counts stay public, as at a real table)."
   },
   "PortModel": {
    "properties": {
     "a": {
      "$ref": "#/components/schemas/CubeModel"
     },
     "b": {
      "$ref": "#/components/schemas/CubeModel"
     },
     "resource": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "sheep",
         "wheat",
         "wood",
         "brick",
         "ore"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Resource"
     }
    },
    "type": "object",
    "required": [
     "a",
     "b",
     "resource"
    ],
    "title": "PortModel",
    "description": "A harbour spanning two coastal vertices.\n\n``resource`` is the resource of a 2:1 port, or ``None`` for a 3:1 general\nport."
   },
   "ReplayStateModel": {
    "properties": {
     "move": {
      "type": "integer",
      "title": "Move"
     },
     "n_moves": {
      "type": "integer",
      "title": "N Moves"
     },
     "board": {
      "$ref": "#/components/schemas/BoardModel"
     },
     "log": {
      "items": {
       "$ref": "#/components/schemas/LogEntryModel"
      },
      "type": "array",
      "title": "Log",
      "default": []
     },
     "winner": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Winner"
     },
     "seats": {
      "anyOf": [
       {
        "items": {
         "type": "string"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Seats"
     }
    },
    "type": "object",
    "required": [
     "move",
     "n_moves",
     "board"
    ],
    "title": "ReplayStateModel",
    "description": "The Replay view's snapshot after ``move`` of ``n_moves`` moves.\n\n``log`` holds the moves played up to that point (the win line appears only\nat the final move); ``winner`` / ``seats`` describe the whole record."
   },
   "ResourceCounts": {
    "properties": {
     "sheep": {
      "type": "integer",
      "title": "Sheep"
     },
     "wheat": {
      "type": "integer",
      "title": "Wheat"
     },
     "wood": {
      "type": "integer",
      "title": "Wood"
     },
     "brick": {
      "type": "integer",
      "title": "Brick"
     },
     "ore": {
      "type": "integer",
      "title": "Ore"
     }
    },
    "type": "object",
    "required": [
     "sheep",
     "wheat",
     "wood",
     "brick",
     "ore"
    ],
    "title": "ResourceCounts",
    "description": "Resource cards in hand, by type (Tile resource order)."
   },
   "RoadModel": {
    "properties": {
     "a": {
      "$ref": "#/components/schemas/CubeModel"
     },
     "b": {
      "$ref": "#/components/schemas/CubeModel"
     },
     "player": {
      "type": "integer",
      "title": "Player"
     }
    },
    "type": "object",
    "required": [
     "a",
     "b",
     "player"
    ],
    "title": "RoadModel",
    "description": "A road along an edge between two vertices, owned by a 0-indexed player."
   },
   "Terrain": {
    "type": "string",
    "enum": [
     "wheat",
     "sheep",
     "wood",
     "ore",
     "brick",
     "desert"
    ],
    "title": "Terrain"
   },
   "TileModel": {
    "properties": {
     "q": {
      "type": "integer",
      "title": "Q"
     },
     "r": {
      "type": "integer",
      "title": "R"
     },
     "terrain": {
      "$ref": "#/components/schemas/Terrain"
     },
     "number": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Number"
     }
    },
    "type": "object",
    "required": [
     "q",
     "r",
     "terrain"
    ],
    "title": "TileModel"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "OAuth2PasswordBearer": {
    "type": "oauth2",
    "flows": {
     "password": {
      "scopes": {},
      "tokenUrl": "api/auth/login"
     }
    }
   }
  }
 }
}