# packages/settlrl-game/src/settlrl_game/models.py
from enum import StrEnum
from typing import Literal

from pydantic import BaseModel


class Terrain(StrEnum):
    wheat = "wheat"
    sheep = "sheep"
    wood = "wood"
    ore = "ore"
    brick = "brick"
    desert = "desert"


class HexModel(BaseModel):
    """An axial (pointy-top) tile coordinate."""

    q: int
    r: int


class TileModel(BaseModel):
    q: int
    r: int
    terrain: Terrain
    number: int | None = None


class CubeModel(BaseModel):
    """A board vertex, in cube coordinates (q + r + s = ±1)."""

    q: int
    r: int
    s: int


class BuildingModel(BaseModel):
    """A settlement or city sitting on a vertex, owned by a 0-indexed player."""

    cube: CubeModel
    player: int
    kind: Literal["settlement", "city"]


class RoadModel(BaseModel):
    """A road along an edge between two vertices, owned by a 0-indexed player."""

    a: CubeModel
    b: CubeModel
    player: int


class PortModel(BaseModel):
    """A harbour spanning two coastal vertices.

    ``resource`` is the resource of a 2:1 port, or ``None`` for a 3:1 general
    port.
    """

    a: CubeModel
    b: CubeModel
    resource: Literal["sheep", "wheat", "wood", "brick", "ore"] | None


class ResourceCounts(BaseModel):
    """Resource cards in hand, by type (Tile resource order)."""

    sheep: int
    wheat: int
    wood: int
    brick: int
    ore: int


class DevCardCounts(BaseModel):
    """Unplayed development cards in hand, by type (DevCard order)."""

    knight: int
    road_building: int
    year_of_plenty: int
    monopoly: int
    victory_point: int


class PlayerModel(BaseModel):
    """Summary stats for one (0-indexed) player.

    The per-type breakdowns are private: the server nulls them out for every
    seat the requester doesn't own (counts stay public, as at a real table).
    """

    player: int
    resource_cards: int  # total resource cards in hand
    dev_cards: int  # total unplayed development cards in hand
    victory_points: int  # building victory points (settlement=1, city=2)
    knights_played: int = 0  # cumulative knights played (public)
    longest_road: bool = False  # holds the Longest Road award (+2)
    largest_army: bool = False  # holds the Largest Army award (+2)
    resources: ResourceCounts | None = None
    dev_card_types: DevCardCounts | None = None


class BankModel(BaseModel):
    """Cards left in the supply: resource stacks plus the development deck."""

    resources: ResourceCounts
    dev_cards: int


class BoardModel(BaseModel):
    tiles: list[TileModel]
    buildings: list[BuildingModel] = []
    roads: list[RoadModel] = []
    ports: list[PortModel] = []
    players: list[PlayerModel] = []
    # Tile the robber currently occupies (axial coordinate), if any.
    robber: HexModel | None = None
    bank: BankModel | None = None


class EdgeModel(BaseModel):
    """A board edge between two vertices (in cube coordinates)."""

    a: CubeModel
    b: CubeModel


class ActionModel(BaseModel):
    """One legal move for the acting player.

    ``flat`` is the game's flat action index (post it back to apply the
    move). ``type`` is the lowercased action type. Depending on the type, at
    most one geometry/resource group below is populated; the rest stay ``None``.
    """

    flat: int
    type: str
    label: str
    # Placement target: a vertex (settlement/city), an edge (road), or a tile
    # (robber / knight, with the optional victim player to steal from).
    vertex: CubeModel | None = None
    edge: EdgeModel | None = None
    tile: HexModel | None = None
    victim: int | None = None
    # Resource choices: monopoly (one), year-of-plenty (two), maritime trade,
    # and the domestic trade proposal (give/receive plus the proposed-to partner).
    resource: str | None = None
    resources: list[str] | None = None
    give: str | None = None
    receive: str | None = None
    partner: int | None = None


class TradeOfferModel(BaseModel):
    """The pending 1:1 domestic trade awaiting the partner's answer."""

    proposer: int
    partner: int
    give: str
    receive: str


class GameStatusModel(BaseModel):
    """Turn-flow snapshot for the live game."""

    phase: str
    current_player: int
    acting_player: int
    dice_roll: int
    has_rolled: bool
    your_turn: bool
    terminal: bool
    winner: int | None = None
    # What controls each seat: "human" or a bot kind (settlrl-agents policy name).
    seats: list[str] = []
    # Set while a proposed trade awaits the partner (phase trade_response).
    trade: TradeOfferModel | None = None


class BotMoveModel(BaseModel):
    """One just-played bot move: the seat that acted and the decoded action."""

    player: int
    action: ActionModel


class LogEntryModel(BaseModel):
    """One line of the game's chat / log.

    ``player`` is the seat the line belongs to (``None``: a spectator's chat
    message). Move entries carry the ``action_type`` (the client maps it to an
    icon) and a short text (the action label, or ``"rolled N"`` for rolls).
    """

    id: int
    kind: Literal["move", "chat", "win"]
    player: int | None = None
    action_type: str | None = None
    text: str = ""


class ReplayStateModel(BaseModel):
    """The Replay view's snapshot after ``move`` of ``n_moves`` moves.

    ``log`` holds the moves played up to that point (the win line appears only
    at the final move); ``winner`` / ``seats`` describe the whole record.
    """

    move: int
    n_moves: int
    board: BoardModel
    log: list[LogEntryModel] = []
    winner: int | None = None
    seats: list[str] | None = None


class PlayerBeliefModel(BaseModel):
    """The observer's proven bounds on one player's hand (``lo == hi``: exact)."""

    player: int
    res_lo: ResourceCounts
    res_hi: ResourceCounts


class BeliefModel(BaseModel):
    """Card counting from one human seat's perspective.

    Everything here is derivable from public information (the card-counting
    tracker), so showing it to that seat never leaks hidden state. The
    observer's own row is omitted — their hand is already on screen.
    """

    observer: int
    players: list[PlayerBeliefModel]


class GameModel(BaseModel):
    """One requester's snapshot: board + status + their legal moves.

    Views are per-seat: ``your_turn`` / ``actions`` / unredacted hands /
    ``belief`` follow the seats the request's tokens prove.
    """

    id: str
    # Change counter: clients drop snapshots older than the one shown.
    version: int = 0
    board: BoardModel
    status: GameStatusModel
    actions: list[ActionModel] = []
    # The last bot move played (cleared when a human acts), for animation.
    bot_move: BotMoveModel | None = None
    # The game's chat / log (moves, chat messages, the win), oldest first.
    log: list[LogEntryModel] = []
    # Card counting for the requester's hand seat; None for spectators.
    belief: BeliefModel | None = None
    # Which seats are claimed (join offers the rest).
    seats_claimed: list[int] = []
    # The seats this requester owns — by seat token or by signed-in account — so
    # an account holder is recognised on any device without the seat token.
    your_seats: list[int] = []